Rename `conftest` to `helpers`
- File only contains a single response json, so it doesn't follow PyTest conventions to use conftest.py for shared fixtures.
